refactor PendingContentEvent
