+++ src/system/BlocksModule/Collectible/PendingContentCollectible.php
@@ -11,7 +11,7 @@ declare(strict_types=1);
  * file that was distributed with this source code.
  */
 
-namespace Zikula\Bundle\CoreBundle\Collection\Collectible;
+namespace Zikula\BlocksModule\Collectible;
 
 /**
  * Class PendingContentCollectible
