+++ src/system/UsersModule/Listener/PendingContentListener.php
@@ -15,7 +15,8 @@ namespace Zikula\UsersModule\Listener;
 
 use Symfony\Component\EventDispatcher\EventSubscriberInterface;
 use Symfony\Contracts\Translation\TranslatorInterface;
-use Zikula\Bundle\CoreBundle\Collection\Collectible\PendingContentCollectible;
+use Zikula\BlocksModule\Collectible\PendingContentCollectible;
+use Zikula\BlocksModule\Event\PendingContentEvent;
 use Zikula\Bundle\CoreBundle\Collection\Container;
 use Zikula\Bundle\CoreBundle\Event\GenericEvent;
 use Zikula\PermissionsModule\Api\ApiInterface\PermissionApiInterface;
@@ -52,20 +53,19 @@ class PendingContentListener implements EventSubscriberInterface
     public static function getSubscribedEvents()
     {
         return [
-            'get.pending_content' => ['pendingContent']
+            PendingContentEvent::class => ['pendingContent']
         ];
     }
 
     /**
-     * Respond to 'get.pending_content' events with registration requests pending approval.
-     * When a 'get.pending_content' event is fired, the Users module will respond with the
+     * Respond to 'PendingContentEvent' with registration requests pending approval.
+     * When a 'PendingContentEvent' is dispatched, the Users module will respond with the
      * number of users that are pending administrator approval.
      * In accordance with the 'get_pending_content' conventions, the count of pending
      * registrations, along with information necessary to access the detailed list, is
-     * assumed as a {@link PendingContentCollectible} and added to the event
-     * subject's collection.
+     * assumed as a {@link PendingContentCollectible} and added to the event's collection.
      */
-    public function pendingContent(GenericEvent $event): void
+    public function pendingContent(PendingContentEvent $event): void
     {
         if (!$this->permissionApi->hasPermission(UsersConstant::MODNAME . '::', '::', ACCESS_MODERATE)) {
             return;
@@ -79,7 +79,7 @@ class PendingContentListener implements EventSubscriberInterface
         if (!empty($numPendingApproval)) {
             $collection = new Container(UsersConstant::MODNAME);
             $collection->add(new PendingContentCollectible('user_registrations', $this->translator->trans('Users pending approval'), $numPendingApproval, 'zikulausersmodule_useradministration_list'));
-            $event->getSubject()->add($collection);
+            $event->addCollection($collection);
         }
     }
 }
